# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Product Service",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/product/create": {
            "post": {
                "description": "CreateNewProduct with json struct",
                "tags": [
                    "Product"
                ],
                "summary": "Creating New Product",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product_service.NewProductDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/products.ProductCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    }
                }
            }
        },
        "/api/product/delete": {
            "delete": {
                "description": "Delete product by UUID",
                "tags": [
                    "Product"
                ],
                "summary": "Deletion of product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of product",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/products.ProductDeleteResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    }
                }
            }
        },
        "/api/product/update": {
            "patch": {
                "description": "Update any field / fields of product\nOnly ID is required",
                "tags": [
                    "Product"
                ],
                "summary": "Updating product information",
                "parameters": [
                    {
                        "description": "Product",
                        "name": "product",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/product_service.ProductUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/products.ProductCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "description": "Getting a list of user products",
                "tags": [
                    "Product"
                ],
                "summary": "Getting a list of user products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/products.ProductList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    }
                }
            }
        },
        "/api/{category_api_name}/products": {
            "get": {
                "description": "Getting a list of user products based on category",
                "tags": [
                    "Product"
                ],
                "summary": "Getting a list of user products based on category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category",
                        "name": "category_api_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/products.ProductList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "product_service.NewProductDTO": {
            "type": "object",
            "properties": {
                "api_name": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "product_service.ProductUpdateDTO": {
            "type": "object",
            "properties": {
                "api_name": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "product_service.UserProduct": {
            "type": "object",
            "properties": {
                "api_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "products.ProductCreateResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "products.ProductDeleteResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "products.ProductList": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product_service.UserProduct"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}